feat: Add bulkCreate endpoint to UserController
This commit adds a new endpoint `bulkCreate` to the `UserController` in the `http-gateway` application. The endpoint allows for the creation of multiple users in a single request by accepting an array of `createUserDto` objects. Each user is created using the `userService.create` method, and the resulting user objects are returned in an array.

The purpose of this change is to provide a more efficient way to create multiple users at once, reducing the number of individual requests required.

Note: This commit message follows the conventional format of "feat: <description>".