refactor: Update project entity with status column
This commit updates the `ProjectEntity` class in the `tcp-project` application. It adds a new `status` column to the entity, which represents the status of the project. The `status` column is of type `enum` and has default value `inactive`.

The purpose of this change is to enhance the project management functionality by allowing the tracking of project statuses.