feat: Add createFile and findAllFile endpoints to DeliveryController
This commit adds the `createFile` and `findAllFile` endpoints to the `DeliveryController` in the `http-gateway` application. The `createFile` endpoint allows for the creation of a new delivery file, associating it with a specific delivery. The `findAllFile` endpoint retrieves all files associated with a given delivery.

The purpose of this change is to enhance the functionality of managing delivery files in the application.

Note: This commit message follows the conventional format of "feat: <description>".