feat: Add @faker-js/faker dependency and delivery cell components
This commit adds the "@faker-js/faker" dependency to the "react-admin" application's package.json file. It also introduces new delivery cell components, including DeliveryRenderAccessCell, DeliveryRenderActionCell, DeliveryRenderFileUrlCell, and DeliveryRenderTagsCell.

The purpose of this change is to enhance the functionality of the application by providing access to fake data generation with the "@faker-js/faker" library and improving the rendering of delivery cells.