feat: Update TcpModule register method to include global option
This commit updates the `register` method in the `TcpModule` of the `tcp-service` package. It adds support for the `global` option, allowing the module to be registered as a global module if set to `true`. By default, the `global` option is set to `false`.

The purpose of this change is to provide more flexibility in registering the `TcpModule` and allow it to be used as a global module when needed.