refactor: Add createFile endpoint to AssetController
This commit adds a new `createFile` endpoint to the `AssetController` in the `http-gateway` application. The endpoint allows for the creation of a new asset file by providing the asset ID and the necessary file details. This change enhances the functionality of the `AssetController` and enables the creation of asset files through the API.